Refactor code and create RestAPI client so that spock integation test and uat can use the common code to call deploydb locally or from another host  - Fixes #163
